+++ storties-search/src/main/java/org/example/stortiessearch/data/persistence/model/PostLikeEntity.java
@@ -1,4 +1,4 @@
-package org.example.stortiessearch.infrastructure.persistence.model;
+package org.example.stortiessearch.data.persistence.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -23,11 +23,11 @@ import org.hibernate.annotations.OnDeleteAction;
 @Builder
 @Getter
 @Table(
-    name = "tbl_post_view",
+    name = "tbl_post_like",
     uniqueConstraints = {
-        @UniqueConstraint(name = "uk_view_post_id_user_id", columnNames = {"post_id", "user_id"})
+        @UniqueConstraint(name = "uk_like_post_id_user_id", columnNames = {"post_id", "user_id"})
     })
-public class PostViewEntity {
+public class PostLikeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
