+++ storties-search/src/main/java/org/example/stortiessearch/data/persistence/repository/PostLikeJpaRepository.java
@@ -1,0 +1,7 @@
+package org.example.stortiessearch.data.persistence.repository;
+
+import org.example.stortiessearch.data.persistence.model.PostLikeEntity;
+import org.springframework.data.repository.CrudRepository;
+
+public interface PostLikeJpaRepository extends CrudRepository<PostLikeEntity, Long> {
+}
